pyproject.toml
[project]
dependencies=[
    "click>=8.3.1",
    "duckdb>=1.4.2",
    "httpx>=0.28.1",
    "ipython>=9.7.0",
    "numpy>=2.3.5",
    "polars>=1.35.2",
    "pydantic>=2.12.4",
    "python-chess>=1.999",
    "pyvis>=0.3.2",
]
description="Add your description here"
name="rep-grow"
readme="README.md"
requires-python=">=3.12"
version="0.1.0"

    [project.scripts]
    grow ="rep_grow.grow:main"
    prune="rep_grow.prune:main"

[tool.maturin]
module-name    ="rep_grow._core"
python-packages=["rep_grow"]
python-source  ="src"

[tool.uv]
cache-keys=[
    { file="Cargo.lock" },
    { file="Cargo.toml" },
    { file="pyproject.toml" },
    { file="src/**/*.rs" },
]

[build-system]
build-backend="maturin"
requires     =["maturin>=1.0,<2.0"]

[dependency-groups]
dev=[
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
    "ty>=0.0.1a27",
]
